linux-cp: fix multicast route setup with lcp-sync
Type: fix
Fixes: 78458aa5781e

Code which tries to avoid removing multicast routes prematurely or
adding them multiple times causes problems when lcp-sync is enabled. At
the time the decision is made whether this is the first interface
address added, multiple addresses may already configured on the VPP
interface. This causes the route to not be added.

Retain the check which avoids premature removal and remove the one
which tries to avoid adding the route multiple times. Adding the route
more than once is innocuous.

Signed-off-by: Matthew Smith <mgsmith@netgate.com>
Change-Id: I2a29f87db5ba97e847a6e29ad2283386b1f1dc7b